Tyr: Upgrade elaticsearch data
