+++ source/tyr/tests/tests_default_settings.py
@@ -7,3 +7,6 @@ EMAIL_CHECK_MX = False
 
 #Validate the email by connecting to the smtp server, but doesn't send an email
 EMAIL_CHECK_SMTP = False
+
+#Path to the directory where the data sources for autocomplete are stocked
+TYR_AUTOCOMPLETE_DIR = "/tmp/autocomplete"
